item price returning from int to double

+++ src/model/Coke.java
@@ -39,7 +39,7 @@ public class Coke implements BeverageComponent {
                      sugars = result.getInt("sugars");
                     carbohydrates = result.getInt("carbohydrates");
                     calories= result.getInt("calories");
-                    price = result.getInt("price");
+                    price = result.getDouble("price");
                     count = result.getInt("count");
                 }
